feat(renderer): Enhance rendering pipeline with new features and optimizations
- Added OnResize method to RendererBase for better handling of render and screen dimensions.
- Introduced FallbackBlit shader for upscaling and format conversion when MetalFX is disabled.
- Updated direct_light.shader to handle motion vectors for sky pixels more efficiently.
- Improved distance approximation kernel with LOD-only tracing to optimize performance.
- Enhanced exposure shader to include a minimum luminance floor for stability.
- Modified indirect_bounce shader to correctly sample depth and normal textures at full resolution.
- Refined temporal accumulation shader to handle NaN checks and blend weights more effectively.
- Added bilateral upsample shader for improved texture quality during upsampling.
- Created benchmark log to track performance metrics across multiple frames.

+++ src/renderer/kernels/shaders/volumetric.shader
@@ -69,8 +69,8 @@ KERNEL(VolumetricFog)(
 #endif
 
     // Clamp to valid depth coord
-    if (fullResCoord.x >= width * 2) fullResCoord.x = width * 2 - 1;
-    if (fullResCoord.y >= height * 2) fullResCoord.y = height * 2 - 1;
+    if (fullResCoord.x >= (uint)(width * 2)) fullResCoord.x = (uint)(width * 2) - 1;
+    if (fullResCoord.y >= (uint)(height * 2)) fullResCoord.y = (uint)(height * 2) - 1;
 
     float depth = TEX_READ_2D(texDepth, fullResCoord).r;
     float2 uv = (AS_FLOAT2(gid) + 0.5f) / make_float2(width, height);
